refactor: re-add int casting

+++ app/Services/Transactions/Aggregates/Fees/Historical/DayAggregate.php
@@ -18,7 +18,7 @@ final class DayAggregate
     {
         return $this->mergeWithPlaceholders(
             (new RangeAggregate())->aggregate(Carbon::now()->subDay()->addHour(), Carbon::now(), 'H'),
-            $this->placeholders(Carbon::now()->subDay()->addHour()->timestamp, Carbon::now()->timestamp, 3600, 'H')->take(24)
+            $this->placeholders((int) Carbon::now()->subDay()->addHour()->timestamp, (int) Carbon::now()->timestamp, 3600, 'H')->take(24)
         );
     }
 }
